+++ controllers/agents.js
@@ -231,8 +231,8 @@ router.get('/groups/:group_id/configuration', cache(), function(req, res) {
  * @apiName GetAgentGroupFile
  * @apiGroup Groups
  *
- * @apiParam {String} group_id Group ID.
- * @apiParam {String} file_name Filename
+ * @apiParam {String} [group_id] Group ID.
+ * @apiParam {String} [file_name] Filename
  * @apiParam {String="conf","rootkit_files", "rootkit_trojans", "rcl"} [type] Type of file.
  *
  * @apiDescription Returns the specified file belonging to the group parsed to JSON.
@@ -274,6 +274,7 @@ router.get('/groups/:group_id/files/:filename', cache(), function(req, res) {
  * @apiParam {Number} [limit=500] Maximum number of elements to return.
  * @apiParam {String} [sort] Sorts the collection by a field or fields (separated by comma). Use +/- at the beginning to list in ascending or descending order.
  * @apiParam {String} [search] Looks for elements with the specified string.
+ * @apiParam {String} [hash] Hash algorithm to use to calculate files checksums.
  *
  * @apiDescription Returns the files belonging to the group.
  *
@@ -287,7 +288,7 @@ router.get('/groups/:group_id/files', cache(), function(req, res) {
     req.apicacheGroup = "agents";
 
     var data_request = {'function': '/agents/groups/:group_id/files', 'arguments': {}};
-    var filters = {'offset': 'numbers', 'limit': 'numbers', 'sort':'sort_param', 'search':'search_param'};
+    var filters = {'offset': 'numbers', 'limit': 'numbers', 'sort':'sort_param', 'search':'search_param', 'hash':'names'};
 
     if (!filter.check(req.query, filters, req, res))  // Filter with error
         return;
@@ -300,6 +301,8 @@ router.get('/groups/:group_id/files', cache(), function(req, res) {
         data_request['arguments']['sort'] = filter.sort_param_to_json(req.query.sort);
     if ('search' in req.query)
         data_request['arguments']['search'] = filter.search_param_to_json(req.query.search);
+    if ('hash' in req.query)
+        data_request['arguments']['hash_algorithm'] = req.query.hash;
 
     if (!filter.check(req.params, {'group_id':'names'}, req, res))  // Filter with error
         return;